backend/ocr/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
paddlepaddle==2.5.2
paddleocr==2.7.0.3
opencv-python==4.6.0.66
pillow==10.1.0
numpy==1.24.3
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI & OpenAI Integration
openai==1.3.0

# Price Intelligence & B2B Platform
# psycopg2-binary==2.9.9
playwright==1.40.0
fake-useragent==1.4.0
aiohttp==3.9.1
pandas==2.1.4
asyncio-throttle==1.0.2

# Database & Analytics
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0

# Monitoring & Observability
prometheus-client==0.19.0
structlog==23.2.0 